Complete homework the silly way. Next time try Enum like TimeUnit.

+++ src/Foot.java
@@ -6,7 +6,7 @@ public class Foot extends Length {
         super(value);
     }
 
-    protected float toInch() {
+    protected double toInch() {
         return value * INCHES_IN_A_FOOT;
     }
 }
